backend/requirements.txt
flask
beautifulsoup4
python-dotenv
redis


pinecone-client
openai
pandas
tiktoken==0.1.1
transformers
torch
langchain